dimg now works! It drags-and-drops with buttery smoothness.
I have greatly streamlined the drag event binder
in tweaks.js- it now takes an array of callbacks:

.drag(
{   pre:	function(e){return e.shiftKey;},
    remember:	function(e){return // remembered data from the drag start
    during:	function(e){
			$(e.currentTarget).offset(
}
)//.drag

remember() is new - it generalizes storing
data from the start of the drag. Whatever
remember() returns is stored and is later
available within other callbacks through:
	this.start_data

This is possible because js will not garbage
collect the callbacks argument since they are
bound to the document, so it can be modified
and accessed later. (A simpler approach would
be to store the data as a global or a temporary
member of $ - but that's an unnecessary evil.)

Nevertheless, it could still use some work.
I need to resolve ownership of a drag when
multiple bindings are in play.
For example, if I start a drag holding the
shift key and release it midway through,
which end() is called on mouseup?

Scaling and cropping are next to implement.
Take care of those before messing with .drag()
anymore- it'll give some immediate distinction
to the different drags.

That's all for now!