multiple fingerframes working
I discovered a glitch in $('div').drag(
this
did not refer to the bound element which owns the
drag, but to the original selector used in binding.
As a result, when one fingerframe was repositioned,
it brought all the others with it.
I have since repaired this.
				heh heh heh...

The HTML editor works very nicely now. The only
major work left is:
	1.   segregate the nicEdit and other panels
		from the HTML editing area, possibly
		with a hard <frame>
	2.   add a binding for adjusting DOM precedence
	3.   add a binding to lock a finished fingerframe
	4.   implement popAroundMenus in earnest so that
		everything else can be tweaked efficiently.
		(Borders, backgrounds, etc.)

Less seriously, I have more ideas for a better name:
{finger,wendig,biegbar,gleit}{frame,rand,fach,feld}
of which my current favorite is gleitfach