editor mode prototype
I have implemented editor modes for translate, scale
and a few others.
Events are bound to a class, rather than any element:
$(document).on('mouseenter',
               '.gleitfach_mode_translate',
               function(){...

Mode classes like 'gleitfach_mode_translate' are
given to/taken from the relevant elements whenever
the editor mode is switched.

Using pure CSS rules to control bindings is more
readable and easier to code. (No more lengthy
filtering 'if' statements for gleitfachDragMode
and similar nonsense)

Incidentally, I had been using the gleitfach object
to hold jQuery selectors:
( gleitfach.selected = '.gleitfach_selected' )
However, this is nonsense. It's an unmotivated
waste of characters and if it becomes necessary
in the future it can easily be grafted back in.

The next step is to organize/conform the rest of
my existing code to this new standard, then add
the features I missed before