CSS editing brainstorm
I have taken a long break from this project, but recently
did a lot of hacking with CSS editing in mind. Among other
things, I encountered MutationObserver, which will be
invaluable on many more fronts than this one.

test_css/ is the result of this directed flailing, and holds
the seed of what I plan to incorporate into gleifach.
A quick summary of what I've learned:
-retrieve the entire DOM's parsed CSS rules
-find the ones which apply to the element of interest
-retrieve those rules' unparsed text (for editing and commenting)
-use a MutationObserver to update the text based on client input

The proof of concept is simple: a contenteditable <div/> with which
a client may edit the <div/>'s own style.

I have no idea how efficient/compatible these methods are-
I'll probably have revise to them a lot to find the fastest and
least collision-prone solution.